backend/requirements.txt
# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Data validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Computer Vision - Local GPU
torch==2.2.0
torchvision==0.17.0
ultralytics==8.1.0  # YOLOv8
easyocr==1.7.1
pillow==10.2.0
opencv-python==4.9.0.80

# LLM - Local GPU via Ollama
ollama==0.1.6

# CUDA utilities
nvidia-ml-py3==7.352.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
numpy==1.26.4

# Development
black==24.2.0
isort==5.13.2
mypy==1.8.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx<0.26.0,>=0.25.2
hypothesis==6.98.0
